Move former DataBlockFactory tests
